MSH-21 feat: add interactive main loop skeleton
