+++ Makefile
@@ -20,6 +20,8 @@ LIBFT = $(LIBFT_DIR)/libft.a
 INCLUDES = -I$(INC_DIR) -I$(LIBFT_DIR)
 
 SRCS = $(SRC_DIR)/main.c \
+	$(SRC_DIR)/shell/main_loop.c \
+	$(SRC_DIR)/shell/run_once.c \
 	$(SRC_DIR)/parsing/tokenize_line.c \
 	$(SRC_DIR)/parsing/token_quoted.c \
 	$(SRC_DIR)/parsing/token_utils.c \
@@ -47,7 +49,7 @@ $(LIBFT):
 	$(MAKE) -C $(LIBFT_DIR)
 
 $(NAME): $(LIBFT) $(OBJS)
-	$(CC) $(CFLAGS) $(OBJS) $(LIBFT) -o $(NAME)
+	$(CC) $(CFLAGS) $(OBJS) $(LIBFT) -lreadline -o $(NAME)
 
 $(OBJ_DIR)/%.o: $(SRC_DIR)/%.c
 	@mkdir -p $(@D)
